Removed win32, wayland is now the only supported backend. Expect that hut always has undecorated windows, will provide utils to roll your own. Moved wayland clipboard/drag&drop to async, cause it was not complex enough.. New sampler_params for specialization.

+++ inc/hut/binpacks.hpp
@@ -143,7 +143,7 @@ struct linear1d {
       accumulated_size += blocks_[i].size_;
       iend = i;
     }
-    for (ssize_t i = _index - 1; i >= 0; i--) {
+    for (ssize_t i = ssize_t(_index) - 1; i >= 0; i--) {
       if (blocks_[i].used_)
         break;
       accumulated_size += blocks_[i].size_;
